requirements.txt
# ──────────────────────────────────────────────────────────────────────────────
# Core API framework
# ──────────────────────────────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0

# ──────────────────────────────────────────────────────────────────────────────
# Agentic AI – LangChain ecosystem
# ──────────────────────────────────────────────────────────────────────────────
langchain>=0.1.16
langchain-openai>=0.1.3
langchain-community>=0.0.36
langgraph>=0.0.40
openai>=1.23.0

# ──────────────────────────────────────────────────────────────────────────────
# Machine Learning
# ──────────────────────────────────────────────────────────────────────────────
scikit-learn>=1.4.0
xgboost>=2.0.3
pandas>=2.2.0
numpy>=1.26.0

# ──────────────────────────────────────────────────────────────────────────────
# Synthetic data generation
# ──────────────────────────────────────────────────────────────────────────────
faker>=24.0.0

# ──────────────────────────────────────────────────────────────────────────────
# Utilities
# ──────────────────────────────────────────────────────────────────────────────
python-dotenv>=1.0.1

# ──────────────────────────────────────────────────────────────────────────────
# Testing
# ──────────────────────────────────────────────────────────────────────────────
pytest>=8.1.0
httpx>=0.27.0          # async test client for FastAPI
